+++ rsass/src/sass/functions/selector.rs
@@ -14,7 +14,7 @@ pub fn create_module() -> Scope {
         let mut s = unnamed(s.get_va::<CssSelectorSet>(name!(selectors)))?
             .into_iter();
         if let Some(base) = s.next() {
-            Ok(s.try_fold(base, |base, s| base.append(s))?.into())
+            Ok(s.try_fold(base, |base, s| base.append(&s))?.into())
         } else {
             Err("At least one selector must be passed.")
                 .named(name!(selectors))
@@ -58,7 +58,7 @@ pub fn create_module() -> Scope {
     def!(f, unify(selector1, selector2), |s| {
         let a: CssSelectorSet = s.get(name!(selector1))?;
         let b: CssSelectorSet = s.get(name!(selector2))?;
-        Ok(a.unify(b).into())
+        Ok(a.unify(&b).into())
     });
 
     f